requirements.txt
altgraph==0.17.4
MouseInfo==0.1.3
packaging==25.0
pefile==2023.2.7
psutil==7.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyMsgBox==2.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytweening==1.2.0
pywin32-ctypes==0.2.3
setuptools==80.9.0
